Move SolutionRunner and SolutionCompiler to core-api
Refactoring to remove compile dependency of exercise-types on the
compiler and runner modules (except in test scope).

This has meant moving interfaces to the core-api module, and updating
the dependencies of webapp to include the required modules.

The HelloWorldExercise has an implicit dependency on
ThreadSolutionRunner (which it really shouldn't have), which has been
exposed in the constructor.